+++ gluoncv/model_zoo/rpn/rpn.py
@@ -1,6 +1,7 @@
 """Region Proposal Networks Definition."""
 from __future__ import absolute_import
 
+import numpy as np
 import mxnet as mx
 from mxnet import gluon
 from mxnet import autograd
@@ -56,7 +57,7 @@ class RPN(gluon.HybridBlock):
     """
     def __init__(self, channels, stride, base_size, scales, ratios, alloc_size,
                  clip, nms_thresh, train_pre_nms, train_post_nms,
-                 test_pre_nms, test_post_nms, min_size, **kwargs):
+                 test_pre_nms, test_post_nms, min_size, dtype='float32', **kwargs):
         super(RPN, self).__init__(**kwargs)
         weight_initializer = mx.init.Normal(0.01)
         with self.name_scope():
@@ -65,7 +66,7 @@ class RPN(gluon.HybridBlock):
             anchor_depth = self.anchor_generator.num_depth
             self.region_proposaler = RPNProposal(
                 clip, nms_thresh, train_pre_nms, train_post_nms,
-                test_pre_nms, test_post_nms, min_size, stds=(1., 1., 1., 1.))
+                test_pre_nms, test_post_nms, min_size, stds=(1., 1., 1., 1.), dtype=dtype)
             self.conv1 = nn.HybridSequential()
             self.conv1.add(nn.Conv2D(channels, 3, 1, 1, weight_initializer=weight_initializer))
             self.conv1.add(nn.Activation('relu'))
